update thông báo xóa nhà hàng & update thêm mới nhà hàng

+++ QuanLyTourDuLich/src/com/myclass/gui/Application.java
@@ -4055,10 +4055,21 @@ JButton btnHopDong_TimKiem = new JButton("Tìm");
 		btnNhaHang_Xoa.addActionListener(new ActionListener() {
 			public void actionPerformed(ActionEvent e) {
 				int selectedRow = tblNhaHang.getSelectedRow();
-				String maDiaDiem = (String) tblNhaHang.getValueAt(selectedRow, 0);
-				
-				nhaHangBUS.deleteById(maDiaDiem);
-				nhaHangTblModel.removeRow(selectedRow);
+				if(selectedRow>=0) {
+					int result=JOptionPane.showConfirmDialog(null,"Bạn có chắc muốn xóa thông tin này?","thông báo",JOptionPane.YES_NO_OPTION); 
+					if(result==JOptionPane.YES_OPTION) {
+					String maDiaDiem = (String) tblNhaHang.getValueAt(selectedRow, 0);
+					
+					nhaHangBUS.deleteById(maDiaDiem);
+					nhaHangTblModel.removeRow(selectedRow);
+					}
+					else if(result==JOptionPane.NO_OPTION) {
+						JOptionPane.showMessageDialog(null,"Thông tin này chưa được xóa!");
+					}
+				}
+				if(selectedRow<0 ) {
+					JOptionPane.showMessageDialog(cardQuanLyTaiKhoan, "Bạn chưa chọn trường dữ liệu nào cả");
+				}
 			}
 		});
 		btnNhaHang_Xoa.setFont(new Font("Tahoma", Font.PLAIN, 16));
@@ -4149,7 +4160,7 @@ JButton btnHopDong_TimKiem = new JButton("Tìm");
 				NhaHangDTO dto = new NhaHangDTO();
 				
 				dto.setMaNhaHang(txtAddMaNhaHang.getText());
-				dto.setTenNhaHang(txtAddTenDiaDiem.getText());
+				dto.setTenNhaHang(txtAddTenNhaHang.getText());
 				dto.setDiaChi(txtAddNhaHang_DiaChi.getText());
 				dto.setChiPhiTrenNguoi(Double.valueOf(txtAddNhaHang_ChiPhiTrenNguoi.getText()));
